FRP.Yampa.AffineSpace

Description

Affine space type relation.

Documentation

class (Floating a, VectorSpace v a) => AffineSpace p v a | p -> v, v -> a where

Affine Space type relation.

An affine space is a set (type) p, and an associated vector space v over a field a.

Minimal complete definition

origin, (.+^), (.-.)

Methods

origin :: p

Origin of the affine space.

(.+^) :: p -> v -> p infix 6

Addition of affine point and vector.

(.-^) :: p -> v -> p infix 6

Subtraction of affine point and vector.

(.-.) :: p -> p -> v infix 6

Subtraction of two points in the affine space, giving a vector.

distance :: p -> p -> a

Distance between two points in the affine space, same as the norm of the vector they form (see '(.-.)'.