FRP.Yampa.Delays

Description

Delays

Basic delays

pre :: SF a a

Uninitialized delay operator.

iPre :: a -> SF a a

Initialized delay operator.

old_pre :: SF a a

Uninitialized delay operator (old implementation).

old_iPre :: a -> SF a a

Initialized delay operator (old implementation).

Timed delays

delay :: Time -> a -> SF a a

Delay a signal by a fixed time t, using the second parameter to fill in the initial t seconds.

Variable delay

pause :: b -> SF a Bool -> SF a b -> SF a b

Given a value in an accumulator (b), a predicate signal function (sfC), and a second signal function (sf), pause will produce the accumulator b if sfC input is True, and will transform the signal using sf otherwise. It acts as a pause with an accumulator for the moments when the transformation is paused.