FRP.Yampa.InternalCore

Description

Domain-specific language embedded in Haskell for programming hybrid (mixed discrete-time and continuous-time) systems. Yampa is based on the concepts of Functional Reactive Programming (FRP) and is structured using arrow combinators.

You can find examples, tutorials and documentation on Yampa here:

www.haskell.org/haskellwiki/Yampa

Structuring a hybrid system in Yampa is done based on two main concepts:

Signal Functions: SF. Yampa is based on the concept of Signal Functions, which are functions from a typed input signal to a typed output signal. Conceptually, signals are functions from Time to Value, where time are the real numbers and, computationally, a very dense approximation (Double) is used.

Events: Event. Values that may or may not occur (and would probably occur rarely). It is often used for incoming network messages, mouse clicks, etc. Events are used as values carried by signals.

A complete Yampa system is defined as one Signal Function from some type a to a type b. The execution of this signal transformer with specific input can be accomplished by means of two functions: reactimate (which needs an initialization action, an input sensing action and an actuation/consumer action and executes until explicitly stopped), and react (which executes only one cycle).

Apart from using normal functions and arrow syntax to define SFs, you can also use several combinators. See [#g:4] for basic signals combinators, [#g:11] for ways of switching from one signal transformation to another, and [#g:16] for ways of transforming Event-carrying signals into continuous signals, [#g:19] for ways of delaying signals, and [#g:21] for ways to feed a signal back to the same signal transformer.

Ways to define Event-carrying signals are given in [#g:7], and FRP.Yampa.Event defines events and event-manipulation functions.

Finally, see [#g:26] for sources of randomness (useful in games).

CHANGELOG:

Adds (most) documentation.

New version using GADTs.

ToDo:

Specialize def. of repeatedly. Could have an impact on invaders.

New defs for accs using SFAcc

Make sure opt worked: e.g.

     repeatedly >>> count >>> arr (fmap sqr)

Introduce SFAccHld.

See if possible to unify AccHld wity Acc??? They are so close.

Introduce SScan. BUT KEEP IN MIND: Most if not all opts would have been possible without GADTs???

Look into pairs. At least pairing of SScan ought to be interesting.

Would be nice if we could get rid of first & second with impunity thanks to Id optimizations. That's a clear win, with or without an explicit pair combinator.

delayEventCat is a bit complicated ...

Random ideas:

What if one used rules to optimize - (arr :: SF a ()) to (constant ()) - (arr :: SF a a) to identity But inspection of invader source code seem to indicate that these are not very common cases at all.

It would be nice if it was possible to come up with opt. rules that are invariant of how signal function expressions are parenthesized. Right now, we have e.g. arr f >>> (constant c >>> sf) being optimized to cpAuxA1 f (cpAuxC1 c sf) whereas it clearly should be possible to optimize to just cpAuxC1 c sf What if we didn't use SF' but SFComp :: tfun -> SF' a b -> SF' b c -> SF' a c ???

The transition function would still be optimized in (pretty much) the current way, but it would still be possible to look inside composed signal functions for lost optimization opts. Seems to me this could be done without too much extra effort/no dupl. work. E.g. new cpAux, the general case:

      cpAux sf1 sf2 = SFComp tf sf1 sf2
          where
              tf dt a = (cpAux sf1' sf2', c)
                  where
                      (sf1', b) = (sfTF' sf1) dt a
                      (sf2', c) = (sfTF' sf2) dt b

The ONLY change was changing the constructor from SF' to SFComp and adding sf1 and sf2 to the constructor app.!

An optimized case: cpAuxC1 b sf1 sf2 = SFComp tf sf1 sf2 So cpAuxC1 gets an extra arg, and we change the constructor. But how to exploit without writing 1000s of rules??? Maybe define predicates on SFComp to see if the first or second sf are interesting, and if so, make reassociate and make a recursive call? E.g. we're in the arr case, and the first sf is another arr, so we'd like to combine the two.

It would also be intersting, then, to know when to STOP playing this game, due to the overhead involved.

Why don't we have a SWITCH constructor that indicates that the structure will change, and thus that it is worthwile to keep looking for opt. opportunities, whereas a plain SF' would indicate that things NEVER are going to change, and thus we can just as well give up?

Documentation

module Control.Arrow

Basic definitions

Time

type Time = Double

Time is used both for time intervals (duration), and time w.r.t. some agreed reference point in time.

type DTime = Double

DTime is the time type for lengths of sample intervals. Conceptually, DTime = R+ = { x in R | x > 0 }. Don't assume Time and DTime have the same representation.

Signal Functions

data SF a b

Signal function that transforms a signal carrying values of some type a into a signal carrying values of some type b. You can think of it as (Signal a -> Signal b). A signal is, conceptually, a function from Time to value.

Constructors

SF

Fields

sfTF :: a -> Transition a b

Future Signal Function

data SF' a b where

Constructors

SFArr :: !(DTime -> a -> Transition a b) -> !(FunDesc a b) -> SF' a b

SFSScan :: !(DTime -> a -> Transition a b) -> !(c -> a -> Maybe (c, b)) -> !c -> b -> SF' a b

SFEP :: !(DTime -> Event a -> Transition (Event a) b) -> !(c -> a -> (c, b, b)) -> !c -> b -> SF' (Event a) b

SFCpAXA :: !(DTime -> a -> Transition a d) -> !(FunDesc a b) -> !(SF' b c) -> !(FunDesc c d) -> SF' a d

SF' :: !(DTime -> a -> Transition a b) -> SF' a b

sfTF' :: SF' a b -> DTime -> a -> Transition a b

sfId :: SF' a a

sfConst :: b -> SF' a b

sfArrG :: (a -> b) -> SF' a b

Scanning

sfSScan :: (c -> a -> Maybe (c, b)) -> c -> b -> SF' a b

type Transition a b = (SF' a b, b)

Function descriptions

data FunDesc a b where

Constructors

FDI :: FunDesc a a

FDC :: b -> FunDesc a b

FDE :: (Event a -> b) -> b -> FunDesc (Event a) b

FDG :: (a -> b) -> FunDesc a b

fdFun :: FunDesc a b -> a -> b

Lifting

arrPrim :: (a -> b) -> SF a b

Lifts a pure function into a signal function (applied pointwise).

arrEPrim :: (Event a -> b) -> SF (Event a) b

Lifts a pure function into a signal function applied to events (applied pointwise).

epPrim :: (c -> a -> (c, b, b)) -> c -> b -> SF (Event a) b