FRP.Yampa.Internals

Description

An interface giving access to some of the internal details of the Yampa implementation.

This interface is indended to be used when the need arises to break abstraction barriers, e.g. for interfacing Yampa to the real world, for debugging purposes, or the like. Be aware that the internal details may change. Relying on this interface means that your code is not insulated against such changes.

Documentation

data Event a

A single possible event occurrence, that is, a value that may or may not occur. Events are used to represent values that are not produced continuously, such as mouse clicks (only produced when the mouse is clicked, as opposed to mouse positions, which are always defined).

Constructors

NoEvent

Event a