FRP.Yampa.Loop

Description

Well-initialised loops

State keeping combinators

Loops with guaranteed well-defined feedback

loopPre :: c -> SF (a, c) (b, c) -> SF a b

Loop with an initial value for the signal being fed back.

loopIntegral :: VectorSpace c s => SF (a, c) (b, c) -> SF a b

Loop by integrating the second value in the pair and feeding the result back. Because the integral at time 0 is zero, this is always well defined.