FRP.Yampa.Miscellany

Description

Collection of entities that really should be part of Haskell base, other packages, or simply have no better home.

Reverse function composition

(#) :: (a -> b) -> (b -> c) -> a -> c infixl 9

Deprecated: Use Control.Arrow.(>>>) and Control.Arrow.(<<<).

Reverse function composition

Arrow plumbing aids

dup :: a -> (a, a)

Duplicate an input.

Maps over lists of pairs

mapFst :: (a -> b) -> [(a, c)] -> [(b, c)]

Deprecated: mapFst is not used by Yampa and will be removed from the next release

Map a function over the first component of pairs in a list.

mapSnd :: (a -> b) -> [(c, a)] -> [(c, b)]

Deprecated: mapSnd is not used by Yampa and will be removed from the next release

Map a function over the second component of pairs in a list.

Generalized tuple selectors

sel3_1 :: (a, b, c) -> a

Deprecated: Use the tuple package instead.

Select the first component of a triple.

sel3_2 :: (a, b, c) -> b

Deprecated: Use the tuple package instead.

Select the second component of a triple.

sel3_3 :: (a, b, c) -> c

Deprecated: Use the tuple package instead.

Select the third component of a triple.

sel4_1 :: (a, b, c, d) -> a

Deprecated: Use the tuple package instead.

Select the first component of a 4-element tuple.

sel4_2 :: (a, b, c, d) -> b

Deprecated: Use the tuple package instead.

Select the second component of a 4-element tuple.

sel4_3 :: (a, b, c, d) -> c

Deprecated: Use the tuple package instead.

Select the third component of a 4-element tuple.

sel4_4 :: (a, b, c, d) -> d

Deprecated: Use the tuple package instead.

Select the fourth component of a 4-element tuple.

sel5_1 :: (a, b, c, d, e) -> a

Deprecated: Use the tuple package instead.

Select the first component of a 5-element tuple.

sel5_2 :: (a, b, c, d, e) -> b

Deprecated: Use the tuple package instead.

Select the second component of a 5-element tuple.

sel5_3 :: (a, b, c, d, e) -> c

Deprecated: Use the tuple package instead.

Select the third component of a 5-element tuple.

sel5_4 :: (a, b, c, d, e) -> d

Deprecated: Use the tuple package instead.

Select the fourth component of a 5-element tuple.

sel5_5 :: (a, b, c, d, e) -> e

Deprecated: Use the tuple package instead.

Select the fifth component of a 5-element tuple.

Floating point utilities

fDiv :: RealFrac a => a -> a -> Integer infixl 7

Deprecated: These are not used by Yampa and will be removed.

Floating-point integer division.

fMod :: RealFrac a => a -> a -> a infixl 7

Deprecated: These are not used by Yampa and will be removed.

Floating-point modulo.

fDivMod :: RealFrac a => a -> a -> (Integer, a)

Deprecated: These are not used by Yampa and will be removed.

Floating-point integer division and modulo.

Liftings

arr2 :: Arrow a => (b -> c -> d) -> a (b, c) d

Lift a binary function onto an arrow

arr3 :: Arrow a => (b -> c -> d -> e) -> a (b, c, d) e

Lift a 3-ary function onto an arrow

arr4 :: Arrow a => (b -> c -> d -> e -> f) -> a (b, c, d, e) f

Lift a 4-ary function onto an arrow

arr5 :: Arrow a => (b -> c -> d -> e -> f -> g) -> a (b, c, d, e, f) g

Lift a 5-ary function onto an arrow

lift0 :: Arrow a => c -> a b c

Lift an 0-ary function onto an arrow

If there was an arr0 function, this would be a synonym.

lift1 :: Arrow a => (c -> d) -> a b c -> a b d

Lift a function into a function between arrows.

lift2 :: Arrow a => (c -> d -> e) -> a b c -> a b d -> a b e

Lift a binary function into a function between arrows.

lift3 :: Arrow a => (c -> d -> e -> f) -> a b c -> a b d -> a b e -> a b f

Lift a 3-ary function into a function between arrows.

lift4 :: Arrow a => (c -> d -> e -> f -> g) -> a b c -> a b d -> a b e -> a b f -> a b g

Lift a 4-ary function into a function between arrows.

lift5 :: Arrow a => (c -> d -> e -> f -> g -> h) -> a b c -> a b d -> a b e -> a b f -> a b g -> a b h

Lift a 5-ary function into a function between arrows.