FRP.Yampa.Random

Description

Documentation

class RandomGen g where

Methods

next :: g -> (Int, g)

genRange :: g -> (Int, Int)

split :: g -> (g, g)

class Random a where

Methods

randomR :: RandomGen g => (a, a) -> g -> (a, g)

random :: RandomGen g => g -> (a, g)

randomRs :: RandomGen g => (a, a) -> g -> [a]

randoms :: RandomGen g => g -> [a]

randomRIO :: (a, a) -> IO a

randomIO :: IO a

Noise (random signal) sources and stochastic event sources

noise :: (RandomGen g, Random b) => g -> SF a b

Noise (random signal) with default range for type in question; based on randoms.

noiseR :: (RandomGen g, Random b) => (b, b) -> g -> SF a b

Noise (random signal) with specified range; based on randomRs.

occasionally :: RandomGen g => g -> Time -> b -> SF a (Event b)

Stochastic event source with events occurring on average once every t_avg seconds. However, no more than one event results from any one sampling interval in the case of relatively sparse sampling, thus avoiding an event backlog should sampling become more frequent at some later point in time.