FRP.Yampa.Core

Description

Minimal FRP core.

For documentation purposes only, to serve as a minimal FRP implementation. Based on Antony Courtney's thesis "Modeling User Interfaces in a Functional Language", page 48 (see https://www.antonycourtney.com/pubs/ac-thesis.pdf, page 61).

Notes:

While time is defined as "core", it is not a primitive in Yampa, and it is actually defined as the integral of 1 over time.

This does not include derivative.

This does not include parallel switching combinators (see Switches).

Signal function

data SF a b

Signal function that transforms a signal carrying values of some type a into a signal carrying values of some type b. You can think of it as (Signal a -> Signal b). A signal is, conceptually, a function from Time to value.

Stateless combinators

iPre :: a -> SF a a

Initialized delay operator.

Creates an SF that delays the input signal, introducing an infinitesimal delay (one sample), using the given argument to fill in the initial output at time zero.

arr :: Arrow a => (b -> c) -> a b c

Lift a function to an arrow.

(>>>) :: Category cat => cat a b -> cat b c -> cat a c infixr 1

Left-to-right composition

first :: Arrow a => a b c -> a (b, d) (c, d)

Send the first component of the input through the argument arrow, and copy the rest unchanged to the output.

Stateful combinators

loop :: ArrowLoop a => a (b, d) (c, d) -> a b c

Instantly loops an SF, making the second output also the second input, using the fix combinator. This introduces a instant loop; without delays, that may lead to an infinite loop.

integral :: VectorSpace a s => SF a a

Integration using the rectangle rule.

Switching upon certain events

data Event a

A single possible event occurrence, that is, a value that may or may not occur. Events are used to represent values that are not produced continuously, such as mouse clicks (only produced when the mouse is clicked, as opposed to mouse positions, which are always defined).

Constructors

NoEvent

Event a

switch :: SF a (b, Event c) -> (c -> SF a b) -> SF a b

Basic switch.

By default, the first signal function is applied. Whenever the second value in the pair actually is an event, the value carried by the event is used to obtain a new signal function to be applied *at that time and at future times*. Until that happens, the first value in the pair is produced in the output signal.

Important note: at the time of switching, the second signal function is applied immediately. If that second SF can also switch at time zero, then a double (nested) switch might take place. If the second SF refers to the first one, the switch might take place infinitely many times and never be resolved.

Remember: The continuation is evaluated strictly at the time of switching!

Time

Note: The function time is actually the integral of 1 over time. So, it's not really necessary.

type Time = Double

Time is used both for time intervals (duration), and time w.r.t. some agreed reference point in time.

time :: SF a Time

Alternative name for localTime.