FRP.Yampa.Delays

Description

SF primitives and combinators to delay signals, introducing new values in them.

Basic delays

pre :: SF a a

Uninitialized delay operator.

The output has an infinitesimal delay (1 sample), and the value at time zero is undefined.

iPre :: a -> SF a a

Initialized delay operator.

Creates an SF that delays the input signal, introducing an infinitesimal delay (one sample), using the given argument to fill in the initial output at time zero.

fby :: b -> SF a b -> SF a b infixr 0

Lucid-Synchrone-like initialized delay (read "followed by").

Initialized delay combinator, introducing an infinitesimal delay (one sample) in given SF, using the given argument to fill in the initial output at time zero.

The difference with iPre is that fby takes an SF as argument.

Timed delays

delay :: Time -> a -> SF a a

Delay a signal by a fixed time t, using the second parameter to fill in the initial t seconds.