FRP.Yampa.Internals

Description

Deprecated: Use FRP.Yampa instead

An interface giving access to some of the internal details of the Yampa implementation.

Legacy, deprecated module.

Documentation

data Event a

A single possible event occurrence, that is, a value that may or may not occur. Events are used to represent values that are not produced continuously, such as mouse clicks (only produced when the mouse is clicked, as opposed to mouse positions, which are always defined).

Constructors

NoEvent

Event a